+++ app/abstract-ccxt-exchange.js
@@ -53,7 +53,7 @@ static getCcxtOpt(exchangeId, config, opt)
     {
         defaultOpt.apiKey = config.exchanges[exchangeId].key;
         defaultOpt.secret = config.exchanges[exchangeId].secret;
-        // needed for Kucoin
+        // needed for Kucoin & Okex
         if (undefined !== config.exchanges[exchangeId].password)
         {
             defaultOpt.password = config.exchanges[exchangeId].password;